+++ packages/ui-components/src/postgrest-infinite-scroll.ts
@@ -17,6 +17,7 @@ const h = hyper.styled(styles);
 interface PostgRESTInfiniteScrollProps extends InfiniteScrollProps<any> {
   id_key: string;
   limit: number;
+  filter_threshold?: number;
   extraParams?: Record<string, any>;
   ascending?: boolean;
   filterable?: boolean;
@@ -40,6 +41,10 @@ interface PostgRESTInfiniteScrollProps extends InfiniteScrollProps<any> {
   }>;
   group_key?: string;
   groups?: Array<{ value: string; label: string }>;
+  GroupingComponent?: React.ComponentType<{
+    title: string;
+    children: React.ReactNode;
+  }>;
 }
 
 export function PostgRESTInfiniteScrollView(
@@ -60,9 +65,11 @@ export function PostgRESTInfiniteScrollView(
     MultiSelectComponent,
     extraParams = {},
     key,
+    GroupingComponent,
     toggles = null,
     searchColumns = undefined,
     group_key = undefined,
+    filter_threshold = 0,
     ...rest
   } = props;
 
@@ -75,6 +82,7 @@ export function PostgRESTInfiniteScrollView(
   );
 
   const [filterValue, setFilterValue] = useState<string>("");
+  const hideData = filterValue.length < filter_threshold;
 
   const SearchBarToUse = SearchBarComponent ?? SearchBar;
   const MultiSelectToUse = MultiSelectComponent ?? MultiSelect;
@@ -279,6 +287,8 @@ export function PostgRESTInfiniteScrollView(
           hasMore: hasMore ?? defaultHasMore,
           key: newKey,
           rest,
+          hideData,
+          GroupingComponent,
         })
       : h(InfiniteScrollView, {
           ...rest,
@@ -288,6 +298,7 @@ export function PostgRESTInfiniteScrollView(
           initialItems: newInitialItems,
           hasMore: hasMore ?? defaultHasMore,
           key: newKey,
+          hideData,
         }),
   ]);
 }
@@ -320,6 +331,8 @@ interface GroupingProps {
   hasMore?: (response: any[]) => boolean;
   key?: string;
   rest?: any;
+  hideData?: boolean;
+  GroupingComponent?: React.ComponentType<GroupingProps>;
 }
 
 function Grouping(props: GroupingProps) {
@@ -332,6 +345,8 @@ function Grouping(props: GroupingProps) {
     getNextParams,
     hasMore,
     rest,
+    hideData,
+    GroupingComponent,
   } = props;
 
   return h("div.group-page", [
@@ -344,6 +359,7 @@ function Grouping(props: GroupingProps) {
         group,
         route,
         id_key,
+        hideData,
         params: {
           ...params,
           [group_key]: "eq." + group.value,
@@ -351,23 +367,36 @@ function Grouping(props: GroupingProps) {
         getNextParams,
         hasMore,
         ...rest,
+        GroupingComponent,
       });
     }),
   ]);
 }
 
 function GroupPanel(props) {
-  const { group, route, params, getNextParams, hasMore, key, ...rest } = props;
+  const {
+    group,
+    route,
+    params,
+    getNextParams,
+    hasMore,
+    key,
+    hideData,
+    GroupingComponent,
+    ...rest
+  } = props;
 
   const data = useAPIResult(route, {
     ...params,
     limit: 1,
   });
 
-  if (!data || data?.length === 0) return null;
+  if (!data || data?.length === 0 || hideData) return null;
+
+  const Panel = GroupingComponent || ExpansionPanel;
 
   return h(
-    ExpansionPanel,
+    Panel,
     {
       title: group.label,
     },
@@ -378,6 +407,7 @@ function GroupPanel(props) {
         params,
         getNextParams,
         hasMore,
+        hideData,
         ...rest,
       }),
     ],